requirements.txt
django==3.2.*
psycopg2-binary==2.9.*
python-dotenv==0.2.*
environs==9.5.*
